requirements.txt
langchain
sentence-transformers
faiss-cpu
streamlit
pypdf
openai
python-dotenv
